+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.8.0
 
 - Add backups storage
+- Update default Kubernetes version to 1.31.3
 
 ## v0.7.0